feature/created components for input fields, added condition rendering for edit button
